requirements.txt
# Core dependencies
crewai>=0.30.0
crewai-tools>=0.2.0
python-dotenv>=1.0.0

# LLM providers
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# Vector database
chromadb>=0.5.0
sentence-transformers>=2.5.0

# News and research APIs
requests>=2.31.0
tavily-python>=0.3.0

# Data models and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
PyYAML>=6.0.1
python-dateutil>=2.8.2

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Optional for better performance
numpy>=1.24.0
pandas>=2.0.0

# Scheduling (for future automation)
schedule>=1.2.0

# GUI dependencies
flet>=0.28.0
